review: reuse common aggregate/output code
