+++ datafusion/physical-plan/src/aggregates/aggregate_hash_table/ordered_partial_table.rs
@@ -29,8 +29,6 @@
 //! The implementation is separated from other aggregate tables because this
 //! execution path is likely to be optimized further in the future.
 
-use std::sync::Arc;
-
 use arrow::datatypes::SchemaRef;
 use arrow::record_batch::RecordBatch;
 use datafusion_common::Result;
@@ -39,7 +37,7 @@ use crate::aggregates::{
     AggregateExec, AggregateMode, aggregate_hash_table::PartialMarker,
 };
 
-use super::common_ordered::{OrderedAggregateTable, remove_emitted_groups};
+use super::common_ordered::OrderedAggregateTable;
 
 /// Implementation specific to partial aggregation, where the table stores
 /// partial aggregate states and the input rows are raw rows.
@@ -75,34 +73,7 @@ impl OrderedAggregateTable<PartialMarker> {
         batch: &RecordBatch,
     ) -> Result<()> {
         let evaluated_batch = self.evaluate_batch(batch)?;
-
-        for group_values in &evaluated_batch.grouping_set_args {
-            let starting_num_groups = self.buffer.group_values.len();
-            self.buffer
-                .group_values
-                .intern(group_values, &mut self.buffer.group_indices)?;
-            let total_num_groups = self.buffer.group_values.len();
-            if total_num_groups > starting_num_groups {
-                self.buffer.group_ordering.new_groups(
-                    group_values,
-                    &self.buffer.group_indices,
-                    total_num_groups,
-                )?;
-            }
-
-            let timer = self.group_by_metrics.aggregation_time.timer();
-            for (acc, values) in self
-                .buffer
-                .accumulators
-                .iter_mut()
-                .zip(evaluated_batch.accumulator_args.iter())
-            {
-                acc.update_batch(values, &self.buffer.group_indices, total_num_groups)?;
-            }
-            drop(timer);
-        }
-
-        Ok(())
+        self.aggregate_evaluated_batch(&evaluated_batch, false)
     }
 
     /// Emits the next batch of partial state rows for groups proven complete by
@@ -122,30 +93,6 @@ impl OrderedAggregateTable<PartialMarker> {
     pub(in crate::aggregates) fn next_output_batch(
         &mut self,
     ) -> Result<Option<RecordBatch>> {
-        if self.buffer.group_values.is_empty() {
-            return Ok(None);
-        }
-
-        let Some(emit_to) = self.buffer.group_ordering.emit_to() else {
-            return Ok(None);
-        };
-        let (emit_to, should_remove_groups) =
-            self.clamp_emit_to(self.buffer.group_values.len(), emit_to);
-
-        let timer = self.group_by_metrics.emitting_time.timer();
-        let mut output = self.buffer.group_values.emit(emit_to)?;
-        if should_remove_groups {
-            remove_emitted_groups(&mut self.buffer.group_ordering, emit_to);
-        }
-
-        for acc in &mut self.buffer.accumulators {
-            output.extend(acc.state(emit_to)?);
-        }
-        drop(timer);
-
-        let batch = RecordBatch::try_new(Arc::clone(&self.output_schema), output)?;
-        debug_assert!(batch.num_rows() > 0);
-
-        Ok(Some(batch))
+        self.next_output_batch_for_mode(false)
     }
 }
